Revert "Feat/inspirations"
