+++ components/Button.tsx
@@ -13,7 +13,6 @@ interface ButtonProps {
 }
 
 const HoverBox = styled(Box)`
-  width: fit-content;
   transition: opacity 0.2s ease-in-out;
   &:active {
     opacity: 0.7;
