+++ components/ProviderList/IntegrationCard.tsx
@@ -1,6 +1,5 @@
 import Box from '@components/Box';
 import IconButton from '@components/IconButton';
-import Spinner from '@components/Spinner';
 import useProviders from '@hooks/useProviders';
 import fetchJson from '@lib/fetchJson';
 import { IntegrationInfos } from '@models/integration';
@@ -21,14 +20,12 @@ interface IntegrationCardProps {
   integration: Integration;
   provider: Provider;
   infos: IntegrationInfos;
-  isStatsLoading: boolean;
 }
 
 export default function IntegrationCard({
   integration,
   provider,
   infos,
-  isStatsLoading,
 }: IntegrationCardProps) {
   const { refresh } = useProviders();
   const { colors } = useTheme();
@@ -117,34 +114,27 @@ export default function IntegrationCard({
         display='flex'
         alignItems='center'
         justifyContent='space-evenly'
-        position='relative'
         py='24px'
         px='8px'
       >
-        {isStatsLoading ? (
-          <Spinner width={20} height={20} icon={false} />
-        ) : (
-          <>
-            <Stat
-              value={infos?.followersCount}
-              label='followers'
-              description='Current total number of followers'
-              Icon={RiUserFollowLine}
-            />
-            <Stat
-              value={infos?.reactionsCount}
-              label='reactions'
-              description='Current total number of reactions on your articles'
-              Icon={MdOutlineAddReaction}
-            />
-            <Stat
-              value={infos?.postsCount}
-              label='posts'
-              description='Current number of article published'
-              Icon={MdPostAdd}
-            />
-          </>
-        )}
+        <Stat
+          value={infos?.followersCount}
+          label='followers'
+          description='Current total number of followers'
+          Icon={RiUserFollowLine}
+        />
+        <Stat
+          value={infos?.reactionsCount}
+          label='reactions'
+          description='Current total number of reactions on your articles'
+          Icon={MdOutlineAddReaction}
+        />
+        <Stat
+          value={infos?.postsCount}
+          label='posts'
+          description='Current number of article published'
+          Icon={MdPostAdd}
+        />
       </Box>
       <Box display='flex' alignItems='center' justifyContent='center' pb='16px'>
         <BsCheckCircleFill size={16} color='#24b47e' />
