--- hooks/useScreenSize.ts
@@ -1,27 +1,0 @@
-import React from 'react';
-
-interface WindowSize {
-  width: number;
-  height: number;
-}
-
-export default function useScreenSize() {
-  const [windowSize, setWindowSize] = React.useState<WindowSize>({
-    width: window.innerWidth,
-    height: window.innerHeight,
-  });
-
-  React.useEffect(() => {
-    window.addEventListener('resize', () => {
-      setWindowSize({ width: window.innerWidth, height: window.innerHeight });
-    });
-
-    return () => {
-      window.removeEventListener('resize', () => {
-        setWindowSize({ width: window.innerWidth, height: window.innerHeight });
-      });
-    };
-  }, []);
-
-  return windowSize;
-}
