--- models/dailyDevItem.ts
@@ -1,7 +1,0 @@
-export interface DailyDevItem {
-  title: string;
-  link: string;
-  guid: string;
-  category: string | string[];
-  pubDate: string;
-}
