+++ package.json
@@ -14,10 +14,8 @@
     "@radix-ui/react-dialog": "^0.1.7",
     "@radix-ui/react-tooltip": "^0.1.7",
     "@sentry/nextjs": "^6.18.2",
-    "@styled-system/css": "^5.1.5",
     "crypto-js": "^4.1.1",
     "date-fns": "^2.28.0",
-    "fast-xml-parser": "^4.0.7",
     "next": "^12.1.0",
     "next-auth": "^4.2.1",
     "next-pwa": "^5.4.5",
