+++ pages/_document.tsx
@@ -61,6 +61,7 @@ class MainDocument extends Document {
         <body>
           <Main />
           <NextScript />
+          <div id='modal-root'></div>
         </body>
       </Html>
     );
