+++ prisma/schema.prisma
@@ -47,7 +47,6 @@ model User {
   email         String?   @unique
   emailVerified DateTime?
   image         String?
-  dailyDevRss   String?
   accounts      Account[]
   sessions      Session[]
   posts        Post[]
