+++ styles/globals.scss
@@ -1,0 +1,52 @@
+@import './colors.scss';
+
+html,
+body {
+  padding: 0;
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Oxygen, Ubuntu,
+    Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+}
+
+body.modal-open {
+  overflow: hidden;
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+.primary {
+  color: $primary;
+}
+
+.pointer {
+  cursor: pointer;
+}
+
+.canvas {
+  background-color: transparent;
+  height: 100%;
+  left: 0px;
+  pointer-events: none;
+  position: fixed;
+  top: 0px;
+  width: 100%;
+}
+
+// custom editor
+.ProseMirror {
+  p {
+    font-size: 1.2rem;
+    line-height: 1.5;
+  }
+
+  h3 {
+    font-size: 1.3rem;
+  }
+}
